refactored checkMeetTime function and rework modulation task
