+++ js/main.js
@@ -1,3 +1,4 @@
 import { getPhotos } from './get-photos.js';
+
 getPhotos();
